requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
tortoise-orm[asyncmy]
python-dotenv
aiomysql
passlib
python-multipart
pyjwt
pydantic